RDKB-59688: [DFS] After triggering a radar, the "nop_started" channel seem not to be passively scanned
Reason for change: Off_chan_scanning should not be performed for nop_started channels.
Test Procedure: 1. Enable DFS and set DFS channels as operating channel.
2. Enable the radar pulse. Shifts to non-DFS channel.
3. Check at /tmp/wifiMon logs whether nop_started channels are listed.
Priority: P1
Risks: Low

+++ source/stats/wifi_stats_radio_channel.c
@@ -775,7 +775,7 @@ int execute_radio_channel_api(wifi_mon_collector_element_t *c_elem, wifi_monitor
     int count = 0;
     int id = 0;
     int on_chan_list[MAX_CHANNELS] = {0};
-    unsigned int nop_chan_list[MAX_DFS_CHANNELS] = { 0 };
+    int nop_chan_list[MAX_DFS_CHANNELS] = { 0 };
     int onchan_num_channels = 0;
     int new_num_channels = 0;
     int updated_channels[MAX_CHANNELS] = {0};
@@ -808,7 +808,7 @@ int execute_radio_channel_api(wifi_mon_collector_element_t *c_elem, wifi_monitor
     }
 
     for (unsigned int j = 0; j < mon_data->nop_channels_num; j++) {
-        nop_chan_list[j] = mon_data->nop_started_channels[j];
+        nop_chan_list[j] = (int)mon_data->nop_started_channels[j];
     }
 
     if (args->scan_mode == WIFI_RADIO_SCAN_MODE_ONCHAN) {
@@ -884,7 +884,7 @@ int execute_radio_channel_api(wifi_mon_collector_element_t *c_elem, wifi_monitor
 
             if (mon_data->nop_channels_num != 0) {
                 for (unsigned int j = 0; j < mon_data->nop_channels_num; j++) {
-                    if (args->channel_list.channels_list[i] == nop_chan_list[j]) {
+                    if ((int)args->channel_list.channels_list[i] == nop_chan_list[j]) {
                         is_nop_chan = 1;
                         break;
                     }
